+++ tests/test_catt.py
@@ -23,9 +23,9 @@ class TestThings(unittest.TestCase):
         self.assertEqual(stream.extractor, "youtube")
 
     def test_stream_info_other_video(self):
-        stream = StreamInfo("https://clips.twitch.tv/CloudyEnticingChickpeaCeilingCat")
+        stream = StreamInfo("https://www.twitch.tv/buddha/clip/OutstandingSquareMallardBudStar")
         self.assertIn("https://", stream.video_url)
-        self.assertEqual(stream.video_id, "304482431")
+        self.assertEqual(stream.video_id, "471296669")
         self.assertTrue(stream.is_remote_file)
         self.assertEqual(stream.extractor, "twitch")